st_nca/federated_pretrain/requirements.txt
pyarrow
networkx
tensordict
torchmetrics
st_nca @ git+https://github.com/petroniocandido/st_nca@main